# iuroadmap.services/api-gateway/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/roadmaps/my": {
      "get": {
        "operationId": "RoadmapsController_getMyRoadmaps",
        "summary": "Get current user roadmap summaries",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User roadmap summaries retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserRoadmapSummaryDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Roadmaps"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/roadmaps/preview/{slug}": {
      "get": {
        "operationId": "RoadmapsController_getPreviewRoadmapBySlug",
        "summary": "Preview a roadmap macro graph by major slug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "description": "Major slug",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Preview roadmap macro graph retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MacroRoadmapResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Roadmaps"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/roadmaps/{userRoadmapId}": {
      "get": {
        "operationId": "RoadmapsController_getMacroRoadmap",
        "summary": "Get macro roadmap details and user progress by user roadmap ID",
        "parameters": [
          {
            "name": "userRoadmapId",
            "required": true,
            "in": "path",
            "description": "User roadmap ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Macro roadmap details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MacroRoadmapResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Roadmaps"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/roadmaps/course-nodes/{courseNodeId}/micro": {
      "get": {
        "operationId": "RoadmapsController_getMicroRoadmap",
        "summary": "Get micro roadmap (topic details and graph) for a course node",
        "parameters": [
          {
            "name": "courseNodeId",
            "required": true,
            "in": "path",
            "description": "Course node ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Micro roadmap retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MicroRoadmapResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Roadmaps"
        ]
      }
    },
    "/api/v1/roadmaps/{userRoadmapId}/courses/{courseNodeId}": {
      "patch": {
        "operationId": "RoadmapsController_updateCourseProgress",
        "summary": "Update student progress (status/credits) on a specific course node",
        "parameters": [
          {
            "name": "userRoadmapId",
            "required": true,
            "in": "path",
            "description": "User roadmap ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "courseNodeId",
            "required": true,
            "in": "path",
            "description": "Course node ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "example": "COMPLETED"
                  },
                  "creditsEarned": {
                    "type": "number",
                    "example": 4
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Course progress updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Roadmaps"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/admin/roadmaps/{roadmapId}/courses/{courseNodeId}": {
      "patch": {
        "operationId": "AdminRoadmapsAliasController_updateCourseNode",
        "summary": "Update an existing course node in a roadmap",
        "parameters": [
          {
            "name": "roadmapId",
            "required": true,
            "in": "path",
            "description": "Roadmap ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "courseNodeId",
            "required": true,
            "in": "path",
            "description": "Course node ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseNodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Course node updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCourseNodeResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin-Roadmaps"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "AdminRoadmapsAliasController_deleteCourseNode",
        "summary": "Delete a course node from a roadmap",
        "parameters": [
          {
            "name": "roadmapId",
            "required": true,
            "in": "path",
            "description": "Roadmap ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "courseNodeId",
            "required": true,
            "in": "path",
            "description": "Course node ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Course node deleted successfully"
          }
        },
        "tags": [
          "Admin-Roadmaps"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "IUROADMAP API Gateway",
    "description": "The API Gateway for the IUROADMAP microservices network.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UserRoadmapSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Roadmap ID",
            "example": 1
          },
          "userId": {
            "type": "number",
            "description": "User ID",
            "example": 1
          },
          "title": {
            "type": "string",
            "description": "Roadmap title",
            "example": "Software Engineering Core"
          },
          "slug": {
            "type": "string",
            "description": "Roadmap slug",
            "example": "software-engineering-core"
          },
          "progressPercent": {
            "type": "number",
            "description": "Progress percentage",
            "example": 45.5
          },
          "totalNodes": {
            "type": "number",
            "description": "Total course nodes",
            "example": 40
          },
          "completedNodes": {
            "type": "number",
            "description": "Completed course nodes",
            "example": 18
          },
          "startDate": {
            "type": "string",
            "description": "Start date",
            "example": "2026-01-15T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "Last updated timestamp",
            "example": "2026-07-11T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "userId",
          "title",
          "slug",
          "progressPercent",
          "totalNodes",
          "completedNodes",
          "startDate",
          "updatedAt"
        ]
      },
      "MacroRoadmapNodeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Node ID",
            "example": 1
          },
          "slug": {
            "type": "string",
            "description": "Node slug",
            "example": "calculus-1"
          },
          "name": {
            "type": "string",
            "description": "Node name",
            "example": "Calculus I"
          },
          "credits": {
            "type": "number",
            "description": "Credits required",
            "example": 3
          },
          "coords": {
            "type": "object",
            "description": "Graph coordinates",
            "example": {
              "x": 0,
              "y": 100
            }
          },
          "status": {
            "type": "string",
            "description": "Node completion status",
            "example": "AVAILABLE",
            "enum": [
              "AVAILABLE",
              "IN_PROGRESS",
              "COMPLETED"
            ]
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "credits",
          "status"
        ]
      },
      "MacroRoadmapEdgeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Edge ID",
            "example": 1
          },
          "from": {
            "type": "number",
            "description": "Source node ID",
            "example": 1
          },
          "to": {
            "type": "number",
            "description": "Target node ID",
            "example": 2
          }
        },
        "required": [
          "id",
          "from",
          "to"
        ]
      },
      "MacroRoadmapResponseDto": {
        "type": "object",
        "properties": {
          "userRoadmapId": {
            "type": "number",
            "description": "User Roadmap ID",
            "example": 1
          },
          "description": {
            "type": "string",
            "description": "Roadmap description",
            "example": "Core engineering curriculum"
          },
          "roadmapId": {
            "type": "number",
            "description": "Roadmap ID",
            "example": 101
          },
          "completion_percentage": {
            "type": "number",
            "description": "Completion percentage",
            "example": 25
          },
          "total_credits_earned": {
            "type": "number",
            "description": "Total credits earned so far",
            "example": 30
          },
          "total_credits_required": {
            "type": "number",
            "description": "Total credits required",
            "example": 130
          },
          "nodes": {
            "description": "Macro roadmap nodes",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MacroRoadmapNodeDto"
            }
          },
          "edges": {
            "description": "Macro roadmap edges",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MacroRoadmapEdgeDto"
            }
          }
        },
        "required": [
          "userRoadmapId",
          "roadmapId",
          "completion_percentage",
          "total_credits_earned",
          "total_credits_required",
          "nodes",
          "edges"
        ]
      },
      "MicroTopicNodeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Topic Node ID",
            "example": 1
          },
          "slug": {
            "type": "string",
            "description": "Topic slug",
            "example": "introduction-to-nestjs"
          },
          "title": {
            "type": "string",
            "description": "Topic title",
            "example": "Introduction to NestJS"
          },
          "description": {
            "type": "string",
            "description": "Topic description",
            "example": "Overview of NestJS concepts"
          },
          "coords": {
            "type": "object",
            "description": "Coordinates for graph rendering",
            "example": {
              "x": 100,
              "y": 200
            }
          },
          "learning_objectives": {
            "type": "string",
            "description": "Learning objectives",
            "example": "Understand Dependency Injection"
          },
          "resources_url": {
            "type": "string",
            "description": "External resources URL",
            "example": "https://docs.nestjs.com"
          }
        },
        "required": [
          "id",
          "slug",
          "title"
        ]
      },
      "MicroTopicEdgeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Edge ID",
            "example": 1
          },
          "from": {
            "type": "number",
            "description": "Source node ID",
            "example": 1
          },
          "to": {
            "type": "number",
            "description": "Target node ID",
            "example": 2
          }
        },
        "required": [
          "id",
          "from",
          "to"
        ]
      },
      "MicroRoadmapResponseDto": {
        "type": "object",
        "properties": {
          "courseNodeId": {
            "type": "number",
            "description": "Course Node ID",
            "example": 10
          },
          "topics": {
            "description": "List of micro topics",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MicroTopicNodeDto"
            }
          },
          "edges": {
            "description": "List of edges connecting micro topics",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MicroTopicEdgeDto"
            }
          }
        },
        "required": [
          "courseNodeId",
          "topics",
          "edges"
        ]
      },
...
      "CoordsDto": {
        "type": "object",
        "properties": {
          "x": {
            "type": "number",
            "description": "X coordinate",
            "example": 100
          },
          "y": {
            "type": "number",
            "description": "Y coordinate",
            "example": 200
          }
        },
        "required": [
          "x",
          "y"
        ]
      },
...
      "AdminCourseNodeResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Course node ID",
            "example": 10
          },
          "roadmap_id": {
            "type": "number",
            "description": "Roadmap ID",
            "example": 1
          },
          "slug": {
            "type": "string",
            "description": "Course node slug",
            "example": "intro-to-programming"
          },
          "name": {
            "type": "string",
            "description": "Course node name",
            "example": "Introduction to Programming"
          },
          "credits": {
            "type": "number",
            "description": "Number of credits",
            "example": 3
          },
          "description": {
            "type": "string",
            "description": "Course description",
            "example": "Basic programming concepts in Python"
          },
          "coords": {
            "type": "object",
            "description": "UI coordinates",
            "example": {
              "x": 100,
              "y": 200
            }
          },
          "created_at": {
            "type": "string",
            "description": "Creation timestamp",
            "example": "2026-07-08T00:00:00.000Z"
          },
          "updated_at": {
            "type": "string",
            "description": "Last update timestamp",
            "example": "2026-07-08T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "roadmap_id",
          "slug",
          "name",
          "credits",
          "created_at",
          "updated_at"
        ]
      },
      "UpdateCourseNodeDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "description": "Course node slug",
            "example": "intro-to-programming"
          },
          "name": {
            "type": "string",
            "description": "Course node name",
            "example": "Introduction to Programming"
          },
          "credits": {
            "type": "number",
            "description": "Number of credits",
            "example": 3
          },
          "description": {
            "type": "string",
            "description": "Course description",
            "example": "Basic programming concepts in Python"
          },
          "coords": {
            "description": "UI coordinates",
            "allOf": [
              {
                "$ref": "#/components/schemas/CoordsDto"
              }
            ]
          },
          "Prerequisites": {
            "description": "Array of prerequisite node IDs",
            "example": [
              1,
              2
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}